src/main/java/org/eventbook/eventbooking/service/impl/MailServiceImpl.java
package org.eventbook.eventbooking.service.impl;

import jakarta.mail.internet.MimeMessage;
import lombok.RequiredArgsConstructor;
import lombok.SneakyThrows;
import org.eventbook.eventbooking.domain.MailType;
import org.eventbook.eventbooking.domain.user.User;
import org.eventbook.eventbooking.service.MailService;
import freemarker.template.Configuration;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.mail.javamail.MimeMessageHelper;
import org.springframework.stereotype.Service;

import java.io.StringWriter;
import java.util.HashMap;
import java.util.Map;
import java.util.Properties;

@Service
@RequiredArgsConstructor
public class MailServiceImpl implements MailService {

    private final Configuration configuration;
    private final JavaMailSender mailSender;
    private final Logger auditlogger =
            LoggerFactory.getLogger("AuditLog");

    @Override
    public void sendEmail(final User user,
                          final MailType type,
                          final Properties params) {
        switch (type) {
            case REGISTRATION -> sendRegistrationEmail(user, params);
            case REMINDER -> sendReminderEmail(user, params);
            default -> { }
        }
    }

    @SneakyThrows
    private void sendRegistrationEmail(final User user,
                                       final Properties params) {
        MimeMessage mimeMessage = mailSender.createMimeMessage();
        MimeMessageHelper helper = new MimeMessageHelper(mimeMessage,
                false,
                "UTF-8");
        helper.setSubject("Thank you for reserving to Event, "
                + user.getName());
        helper.setTo(user.getEmail());
        String emailContent = getRegistrationEmailContent(user, params);
        helper.setText(emailContent, true);
        auditlogger.info("Reserve user to event. User id = {}",
                user.getId());
        mailSender.send(mimeMessage);
    }

    @SneakyThrows
    private void sendReminderEmail(final User user, final Properties params) {
        MimeMessage mimeMessage = mailSender.createMimeMessage();
        MimeMessageHelper helper = new MimeMessageHelper(mimeMessage,
                false,
                "UTF-8");
        helper.setSubject("The event on which you reserve "
                + "tickets will be held tomorrow.");
        helper.setTo(user.getEmail());
        String emailContent = getReminderEmailContent(user, params);
        helper.setText(emailContent, true);
        mailSender.send(mimeMessage);
    }

    @SneakyThrows
    private String getRegistrationEmailContent(final User user,
                                               final Properties properties) {
        StringWriter writer = new StringWriter();
        Map<String, Object> model = new HashMap<>();
        model.put("name", user.getName());
        configuration.getTemplate("register.ftlh")
                .process(model, writer);
        return writer.getBuffer().toString();
    }

    @SneakyThrows
    private String getReminderEmailContent(final User user,
                                           final Properties properties) {
        StringWriter writer = new StringWriter();
        Map<String, Object> model = new HashMap<>();
        model.put("name", user.getName());
        model.put("eventName", properties.getProperty("event.name"));
        model.put("description", properties.getProperty("event.description"));
        model.put("date", properties.getProperty("event.date"));
        configuration.getTemplate("reminder.ftlh")
                .process(model, writer);
        return writer.getBuffer().toString();
    }

}
